e-library-delete- changes

+++ delete.php
@@ -1,15 +1,16 @@
 <?php 
 include 'connection.php';
-$del_id = $_GET['id'];
+if (isset($_GET['id'])) {
+    $del_id = $_GET['id'];
 
 
-$del_book_query = "DELETE FROM addbook WHERE id = '$del_id'";
-$delete_book_result = mysqli_query($con, $del_book_query);
-if($delete_book_result){
+    $del_book_query = "DELETE FROM addbook WHERE id = '$del_id'";
+    $delete_book_result = mysqli_query($con, $del_book_query);
+    if($delete_book_result) {
 
-    header("location:mainpage.php");
+        header("location:mainpage.php");
+    }
 }
-
 
 
 ?>
